requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
gunicorn==23.0.0
flask-login==0.6.3
apify-client
